Refactoring here and there
https://github.com/metasfresh/metasfresh/issues/3856
